+++ CHANGES.md
@@ -4,8 +4,9 @@ Notable changes between versions.
 
 ## Latest
 
+* Kubernetes [v1.26.2](https://github.com/kubernetes/kubernetes/blob/master/CHANGELOG/CHANGELOG-1.26.md#v1262)
 * Update Cilium from v1.12.5 to [v1.12.6](https://github.com/cilium/cilium/releases/tag/v1.12.6)
-* Update flannel from v0.20.2 to [v0.21.1](https://github.com/flannel-io/flannel/releases/tag/v0.21.1)
+* Update flannel from v0.20.2 to [v0.21.2](https://github.com/flannel-io/flannel/releases/tag/v0.21.2)
 
 ### Bare-Metal